Refactor http requests
