+++ src/api/handcash_connect_api_error.ts
@@ -1,13 +1,27 @@
 export default class HandCashConnectApiError extends Error {
+	path: string;
+
+	method: string;
+
 	httpStatusCode: number;
 
 	info: string;
 
-	constructor(httpStatusCode: number, message: string, info: string) {
-		super();
-		this.httpStatusCode = httpStatusCode;
-		this.message = message;
-		this.info = info;
+	constructor(params: {
+		method: string;
+		path: string;
+		httpStatusCode: number;
+		message: string;
+		info: string;
+		stack?: string;
+	}) {
+		super(params.message);
+		this.method = params.method;
+		this.path = params.path;
+		this.httpStatusCode = params.httpStatusCode;
+		this.message = params.message;
+		this.info = params.info;
+		this.stack = params.stack;
 	}
 
 	toString() {
