+++ src/run_extension/handcash_owner.ts
@@ -1,5 +1,4 @@
 import HandCashConnectService from '../api/handcash_connect_service';
-import HttpRequestFactory from '../api/http_request_factory';
 import Environments from '../environments';
 
 export default class HandCashOwner {
@@ -10,15 +9,13 @@ export default class HandCashOwner {
 	}
 
 	static fromAuthToken(authToken: string, env = Environments.prod, appSecret = '', appId = '') {
-		const handCashConnectService = new HandCashConnectService(
-			new HttpRequestFactory({
-				authToken,
-				baseApiEndpoint: env.apiEndpoint,
-				baseTrustholderEndpoint: env.trustholderEndpoint,
-				appSecret,
-				appId,
-			})
-		);
+		const handCashConnectService = new HandCashConnectService({
+			authToken,
+			baseApiEndpoint: env.apiEndpoint,
+			baseTrustholderEndpoint: env.trustholderEndpoint,
+			appSecret,
+			appId,
+		});
 		return new HandCashOwner(handCashConnectService);
 	}
