+++ src/test/unit/handcash_connect.spec.ts
@@ -1,4 +1,4 @@
-import { describe, it, expect } from 'vitest';
+import { describe, expect, it } from 'vitest';
 import Environments from '../../environments';
 import HandCashConnect from '../../handcash_connect';
 
@@ -11,5 +11,17 @@ describe('# HandCashConnect - Unit Tests', () => {
 			appSecret,
 		});
 		expect(handCashConnect.env).to.eq(Environments.prod);
+	});
+
+	it('should raise an invalid auth token error', async () => {
+		const authToken = 'invalid-token-123';
+		const appSecret = '1234567890';
+		const appId = 'id1';
+		return expect(() =>
+			new HandCashConnect({
+				appSecret,
+				appId,
+			}).getAccountFromAuthToken(authToken)
+		).to.throw('Invalid authToken');
 	});
 });
